simplified setup

+++ conf/metadata.php
@@ -3,5 +3,5 @@
  * Options for the twofactoraltemail plugin
  */
 
-$meta['tolerance'] = array('int', '_min' => 2);
+$meta['tolerance'] = array('numeric', '_min' => 2);
